pyproject.toml
[project]
name = "cyber-flowchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.2",
    "psutil>=7.0.0",
    "pythonping>=1.1.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "scapy>=2.6.1",
    "tqdm>=4.67.1",
]
